+++ feature-libs/order/components/order-details/index.ts
@@ -6,6 +6,7 @@
 
 export * from './order-detail-actions/order-detail-actions.component';
 export * from './order-detail-items/index';
+export * from './order-detail-shipping/order-detail-shipping.component';
 export * from './order-detail-totals/order-detail-totals.component';
 export * from './order-details.module';
 export * from './order-details.service';
