+++ feature-libs/order/components/order-details/order-detail-items/order-detail-items.component.html
@@ -20,7 +20,7 @@
     [buyItAgainTranslation]="'addToCart.buyItAgain' | cxTranslate"
   ></cx-order-consigned-entries>
 
-  <!-- delivery info -->
+  <!-- delivery info
   <div class="cx-summary-card">
     <ng-container *ngIf="order.deliveryAddress">
       <cx-card
@@ -34,6 +34,7 @@
       ></cx-card>
     </ng-container>
   </div>
+ -->
 
   <!-- delivery consigned entries -->
   <cx-order-consigned-entries
