+++ feature-libs/order/components/order-details/order-detail-shipping/order-detail-shipping.component.html
@@ -1,0 +1,15 @@
+<ng-container *ngIf="order$ | async as order">
+  <div class="cx-summary-card">
+    <ng-container *ngIf="order.deliveryAddress">
+      <cx-card
+        [content]="getAddressCardContent(order.deliveryAddress) | async"
+      ></cx-card>
+    </ng-container>
+
+    <ng-container *ngIf="order.deliveryMode">
+      <cx-card
+        [content]="getDeliveryModeCardContent(order.deliveryMode) | async"
+      ></cx-card>
+    </ng-container>
+  </div>
+</ng-container>
