+++ feature-libs/order/components/order-details/order-detail-shipping/order-detail-shipping.component.ts
@@ -1,0 +1,72 @@
+/*
+ * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { Component } from '@angular/core';
+import { DeliveryMode } from '@spartacus/cart/base/root';
+import { Address, TranslationService } from '@spartacus/core';
+import { Card } from '@spartacus/storefront';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+import { OrderDetailsService } from '../order-details.service';
+
+@Component({
+  selector: 'cx-order-details-shipping',
+  templateUrl: './order-detail-shipping.component.html',
+})
+export class OrderDetailShippingComponent {
+  order$ = this.orderDetailsService.getOrderDetails();
+
+  constructor(
+    protected orderDetailsService: OrderDetailsService,
+    protected translation: TranslationService
+  ) {}
+
+  getAddressCardContent(deliveryAddress: Address): Observable<Card> {
+    return this.translation.translate('addressCard.shipTo').pipe(
+      filter(() => Boolean(deliveryAddress)),
+      map((textTitle) => {
+        const formattedAddress = this.normalizeFormattedAddress(
+          deliveryAddress.formattedAddress ?? ''
+        );
+
+        return {
+          title: textTitle,
+          textBold: `${deliveryAddress.firstName} ${deliveryAddress.lastName}`,
+          text: [formattedAddress, deliveryAddress.country?.name],
+        } as Card;
+      })
+    );
+  }
+
+  getDeliveryModeCardContent(deliveryMode: DeliveryMode): Observable<Card> {
+    return this.translation.translate('orderDetails.shippingMethod').pipe(
+      filter(() => Boolean(deliveryMode)),
+      map(
+        (textTitle) =>
+          ({
+            title: textTitle,
+            textBold: deliveryMode.name,
+            text: [
+              deliveryMode.description,
+              deliveryMode.deliveryCost?.formattedValue
+                ? deliveryMode.deliveryCost?.formattedValue
+                : '',
+            ],
+          } as Card)
+      )
+    );
+  }
+
+  private normalizeFormattedAddress(formattedAddress: string): string {
+    const addresses = formattedAddress
+      .split(',')
+      .map((address) => address.trim());
+
+    const newFormattedAddress = addresses.filter(Boolean).join(', ');
+
+    return newFormattedAddress;
+  }
+}
