+++ feature-libs/order/components/order-details/order-details.module.ts
@@ -29,6 +29,7 @@ import { TrackingEventsComponent } from './order-detail-items/consignment-tracki
 import { defaultConsignmentTrackingLayoutConfig } from './order-detail-items/default-consignment-tracking-layout.config';
 import { OrderConsignedEntriesComponent } from './order-detail-items/order-consigned-entries/order-consigned-entries.component';
 import { OrderDetailItemsComponent } from './order-detail-items/order-detail-items.component';
+import { OrderDetailShippingComponent } from './order-detail-shipping/order-detail-shipping.component';
 import { OrderDetailTotalsComponent } from './order-detail-totals/order-detail-totals.component';
 import { OrderOverviewComponent } from './order-overview/order-overview.component';
 
@@ -36,6 +37,7 @@ const moduleComponents = [
   OrderDetailActionsComponent,
   OrderDetailItemsComponent,
   OrderDetailTotalsComponent,
+  OrderDetailShippingComponent,
   TrackingEventsComponent,
   ConsignmentTrackingComponent,
   OrderConsignedEntriesComponent,
@@ -70,9 +72,9 @@ const moduleComponents = [
         AccountOrderDetailsTotalsComponent: {
           component: OrderDetailTotalsComponent,
         },
-        // AccountOrderDetailsShippingComponent: {
-        //   component: OrderOverviewComponent,
-        // },
+        AccountOrderDetailsShippingComponent: {
+          component: OrderDetailShippingComponent,
+        },
         AccountOrderDetailsOverviewComponent: {
           component: OrderOverviewComponent,
         },
