+++ feature-libs/order/components/order-details/order-overview/order-overview.component.html
@@ -78,22 +78,6 @@
       </div>
     </ng-container>
 
-    <!--
-    <div class="cx-summary-card">
-      <ng-container *ngIf="order.deliveryAddress">
-        <cx-card
-          [content]="getAddressCardContent(order?.deliveryAddress) | async"
-        ></cx-card>
-      </ng-container>
-
-      <ng-container *ngIf="order.deliveryMode">
-        <cx-card
-          [content]="getDeliveryModeCardContent(order?.deliveryMode) | async"
-        ></cx-card>
-      </ng-container>
-    </div>
-    -->
-
     <ng-container *ngIf="order.paymentInfo">
       <div class="cx-summary-card">
         <cx-card
