+++ feature-libs/order/components/order-details/order-overview/order-overview.component.ts
@@ -156,42 +156,6 @@ export class OrderOverviewComponent {
     );
   }
 
-  // getAddressCardContent(deliveryAddress: Address): Observable<Card> {
-  //   return this.translation.translate('addressCard.shipTo').pipe(
-  //     filter(() => Boolean(deliveryAddress)),
-  //     map((textTitle) => {
-  //       const formattedAddress = this.normalizeFormattedAddress(
-  //         deliveryAddress.formattedAddress ?? ''
-  //       );
-
-  //       return {
-  //         title: textTitle,
-  //         textBold: `${deliveryAddress.firstName} ${deliveryAddress.lastName}`,
-  //         text: [formattedAddress, deliveryAddress.country?.name],
-  //       } as Card;
-  //     })
-  //   );
-  // }
-
-  // getDeliveryModeCardContent(deliveryMode: DeliveryMode): Observable<Card> {
-  //   return this.translation.translate('orderDetails.shippingMethod').pipe(
-  //     filter(() => Boolean(deliveryMode)),
-  //     map(
-  //       (textTitle) =>
-  //         ({
-  //           title: textTitle,
-  //           textBold: deliveryMode.name,
-  //           text: [
-  //             deliveryMode.description,
-  //             deliveryMode.deliveryCost?.formattedValue
-  //               ? deliveryMode.deliveryCost?.formattedValue
-  //               : '',
-  //           ],
-  //         } as Card)
-  //     )
-  //   );
-  // }
-
   getPaymentInfoCardContent(payment: PaymentDetails): Observable<Card> {
     return combineLatest([
       this.translation.translate('paymentForm.payment'),
@@ -228,14 +192,4 @@ export class OrderOverviewComponent {
       )
     );
   }
-
-  // private normalizeFormattedAddress(formattedAddress: string): string {
-  //   const addresses = formattedAddress
-  //     .split(',')
-  //     .map((address) => address.trim());
-
-  //   const newFormattedAddress = addresses.filter(Boolean).join(', ');
-
-  //   return newFormattedAddress;
-  // }
 }
